refactor integration tests to use shared helpers
