+++ tests/integration/alb/test_alb_deprovisioning.py
@@ -1,10 +1,17 @@
 import pytest  # noqa F401
 
 from broker.extensions import db
-from broker.models import Operation, ALBServiceInstance
+from broker.models import ALBServiceInstance
 from tests.lib import factories
 from tests.lib.client import check_last_operation_description
-from tests.lib.alb.deprovision import subtest_deprovision_removes_certificate_from_iam
+from tests.lib.alb.deprovision import (
+    subtest_deprovision_creates_deprovision_operation,
+    subtest_deprovision_removes_ALIAS_records,
+    subtest_deprovision_removes_TXT_records,
+    subtest_deprovision_removes_cert_from_alb,
+    subtest_deprovision_removes_certificate_from_iam,
+    subtest_deprovision_marks_operation_as_succeeded,
+)
 
 
 @pytest.fixture
@@ -72,11 +79,12 @@ def service_instance():
 
 
 def test_deprovision_happy_path(
-    client, service_instance, dns, tasks, route53, iam_govcloud, simple_regex, alb
+    client, service_instance, tasks, route53, iam_govcloud, alb
 ):
-    service_instance = db.session.get(ALBServiceInstance, "1234")
+    instance_model = ALBServiceInstance
+    service_instance = db.session.get(instance_model, "1234")
     operation_id = subtest_deprovision_creates_deprovision_operation(
-        client, service_instance
+        client, service_instance, instance_model
     )
     check_last_operation_description(client, "1234", operation_id, "Queuing tasks")
     subtest_deprovision_removes_ALIAS_records(tasks, route53)
@@ -87,84 +95,17 @@ def test_deprovision_happy_path(
     check_last_operation_description(
         client, "1234", operation_id, "Removing DNS TXT records"
     )
-    subtest_deprovision_removes_cert_from_alb(tasks, service_instance, alb)
+    subtest_deprovision_removes_cert_from_alb(
+        tasks, service_instance, alb, instance_model
+    )
     check_last_operation_description(
         client, "1234", operation_id, "Removing SSL certificate from load balancer"
     )
     subtest_deprovision_removes_certificate_from_iam(
-        tasks, service_instance, iam_govcloud
+        tasks, service_instance, iam_govcloud, instance_model
     )
     check_last_operation_description(
         client, "1234", operation_id, "Removing SSL certificate from AWS"
     )
-    subtest_deprovision_marks_operation_as_succeeded(tasks)
+    subtest_deprovision_marks_operation_as_succeeded(tasks, instance_model)
     check_last_operation_description(client, "1234", operation_id, "Complete!")
-
-
-def subtest_deprovision_creates_deprovision_operation(client, service_instance):
-    service_instance = db.session.get(ALBServiceInstance, "1234")
-    client.deprovision_alb_instance(service_instance.id, accepts_incomplete="true")
-
-    assert client.response.status_code == 202, client.response.body
-    assert "operation" in client.response.json
-
-    operation_id = client.response.json["operation"]
-    operation = db.session.get(Operation, operation_id)
-
-    assert operation is not None
-    assert operation.state == Operation.States.IN_PROGRESS.value
-    assert operation.action == Operation.Actions.DEPROVISION.value
-    assert operation.service_instance_id == service_instance.id
-
-    return operation_id
-
-
-def subtest_deprovision_removes_ALIAS_records(tasks, route53):
-    route53.expect_remove_ALIAS(
-        "example.com.domains.cloud.test", "fake1234.cloud.test", "ALBHOSTEDZONEID"
-    )
-    route53.expect_remove_ALIAS(
-        "foo.com.domains.cloud.test", "fake1234.cloud.test", "ALBHOSTEDZONEID"
-    )
-
-    # one for marking provisioning tasks canceled, which is tested elsewhere
-    tasks.run_queued_tasks_and_enqueue_dependents()
-    tasks.run_queued_tasks_and_enqueue_dependents()
-
-    route53.assert_no_pending_responses()
-
-
-def subtest_deprovision_removes_TXT_records(tasks, route53):
-    route53.expect_remove_TXT(
-        "_acme-challenge.example.com.domains.cloud.test", "example txt"
-    )
-    route53.expect_remove_TXT("_acme-challenge.foo.com.domains.cloud.test", "foo txt")
-
-    tasks.run_queued_tasks_and_enqueue_dependents()
-
-    route53.assert_no_pending_responses()
-
-
-def subtest_deprovision_removes_cert_from_alb(tasks, service_instance, alb):
-    service_instance = db.session.get(ALBServiceInstance, "1234")
-    alb.expect_remove_certificate_from_listener(
-        service_instance.alb_listener_arn,
-        service_instance.current_certificate.iam_server_certificate_arn,
-    )
-    tasks.run_queued_tasks_and_enqueue_dependents()
-    alb.assert_no_pending_responses()
-
-
-def subtest_deprovision_marks_operation_as_succeeded(tasks):
-    db.session.expunge_all()
-    service_instance = db.session.get(ALBServiceInstance, "1234")
-    assert not service_instance.deactivated_at
-
-    tasks.run_queued_tasks_and_enqueue_dependents()
-
-    db.session.expunge_all()
-    service_instance = db.session.get(ALBServiceInstance, "1234")
-    assert service_instance.deactivated_at
-
-    operation = service_instance.operations.first()
-    assert operation.state == "succeeded"
